+++ Creators/CMakeLists.txt
@@ -2,4 +2,6 @@ cmake_minimum_required(VERSION 3.1)
 
 project(Creators)
 
-add_library(Creators ElfCreator.cpp UnitCreator.cpp HealerCreator.cpp WarriorElfCreator.cpp MageCreator.cpp ArchmageCreator.cpp)
+set(SOURCE ElfCreator.cpp UnitCreator.cpp HealerCreator.cpp WarriorElfCreator.cpp MageCreator.cpp ArchmageCreator.cpp)
+
+add_library(Creators ${SOURCE})
