--- Creators/CharactersConfig.h
@@ -1,45 +1,0 @@
-#pragma once
-
-// Attack types
-const std::string RangeAttack = "range";
-const std::string MeleeAttack = "melee";
-
-// Elf stats
-const int ElfHealth = 100;
-const int ElfRegen = 3;
-const int ElfAttackLow = 4;
-const int ElfAttackHigh = 6;
-
-// Healer elf stats
-const int HealerHealth = 120;
-const int HealerRegen = 6;
-const int HealerAttackLow = 4;
-const int HealerAttackHigh = 6;
-const int HealerHealValue = 10;
-const int HealerHealCharges = -1;
-
-// Warrior elf stats
-const int WarriorHealth = 160;
-const int WarriorRegen = 6;
-const int WarriorAttackLow = 6;
-const int WarriorAttackHigh = 10;
-const int WarriorHealValue = 15;
-const int WarriorHealCharges = 3;
-
-// Mage elf stats
-const int MageHealth = 100;
-const int MageRegen = 3;
-const int MageAttackLow = 4;
-const int MageAttackHigh = 6;
-const int MageMagicDmgLow = 12;
-const int MageMagicDmgHigh = 20;
-
-// Archmage elf stats
-const int ArchmageHealth = 120;
-const int ArchmageRegen = 6;
-const int ArchmageAttackLow = 4;
-const int ArchmageAttackHigh = 6;
-const int ArchmageHealValue = 12;
-const int ArchmageHealCharges = -1;
-const int ArchmageMagicDmgLow = 14;
-const int ArchmageMagicDmgHigh = 22;
