+++ Creators/ElfCreator.cpp
@@ -1,5 +1,5 @@
 #include "ElfCreator.h"
-#include "../Units/Elf.h"
+#include "Elf.h"
 #include "CharactersConfig.h"
 
 std::unique_ptr<Elf> ElfCreator::create_unit() const {
