+++ Creators/Headers/ArchmageCreator.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <memory>
+#include "UnitCreator.h"
+#include "Archmage.h"
+
+class ArchmageCreator : public UnitCreator<Archmage> {
+public:
+  std::unique_ptr<Archmage> render_unit();
+  std::unique_ptr<Archmage> create_unit() const override;
+};
