--- Creators/MageCreator.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <memory>
-#include "UnitCreator.h"
-#include "../Units/Mage.h"
-
-class MageCreator : public UnitCreator<Mage> {
-public:
-  std::unique_ptr<Mage> create_unit() const override;
-};
