+++ Units/CMakeLists.txt
@@ -2,4 +2,6 @@ cmake_minimum_required(VERSION 3.1)
 
 project(Units)
 
-add_library(Units Elf.cpp UnitInterface.cpp Healer.cpp WarriorElf.cpp Mage.cpp Archmage.cpp)
+set(SOURCE Elf.cpp UnitInterface.cpp Healer.cpp WarriorElf.cpp Mage.cpp Archmage.cpp)
+
+add_library(Units ${SOURCE})
