+++ Units/Headers/WarriorElf.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Elf.h"
+#include "Healer.h"
+
+class WarriorElf : public Healer {
+  public:
+    std::string quote() const;
+};
